Removed formal languages into separate module
Created a new module called FormalLanguages which contains e0-e5. The Language module is restricted now to the operations needed to create the formal languages for eson. Tests that validate the languages can also be found in a separate test file.
