+++ test/token_seq_test.rb
@@ -40,8 +40,8 @@ end
 describe Eson::Tokenizer::TokenSeq do
 
   before do
-    @alternation_rule = Eson::Language.e1.word_form
-    @concatenation_rule = Eson::Language.e0.variable_identifier
+    @alternation_rule = Eson::FormalLanguages.e1.word_form
+    @concatenation_rule = Eson::FormalLanguages.e0.variable_identifier
     @token_seq = Eson::Tokenizer::TokenSeq.new(4) {Eson::Tokenizer::TokenSeq::Token.new}
   end
